+++ packages/chakra-ui-core/src/Css/index.js
@@ -1,5 +1,2 @@
-import _css from '@styled-system/css'
-import { tx } from '../utils'
-
-const css = styleProps => _css(tx(styleProps))
-export default css
+import Css from './Css'
+export default Css