+++ packages/chakra-ui-docs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/chakra-ui/chakra-ui-vue/compare/chakra-ui-docs@0.2.3...chakra-ui-docs@0.2.4) (2020-03-24)
+
+**Note:** Version bump only for package chakra-ui-docs
+
+
+
+
+
+## [0.2.3](https://github.com/chakra-ui/chakra-ui-vue/compare/chakra-ui-docs@0.2.2...chakra-ui-docs@0.2.3) (2020-03-22)
+
+**Note:** Version bump only for package chakra-ui-docs
+
+
+
+
+
+## [0.2.2](https://github.com/chakra-ui/chakra-ui-vue/compare/chakra-ui-docs@0.2.1...chakra-ui-docs@0.2.2) (2020-03-22)
+
+**Note:** Version bump only for package chakra-ui-docs
+
+
+
+
+
 ## [0.2.1](https://github.com/chakra-ui/chakra-ui-vue/compare/chakra-ui-docs@0.2.0...chakra-ui-docs@0.2.1) (2020-03-21)
 
 **Note:** Version bump only for package chakra-ui-docs